backend/requirements/dev.txt
-r prod.txt

pytest
pytest-django
pytest-cov
factory_boy
coverage
black
ruff
pre-commit
freezegun
ptpython
django-extensions
ipython
bpython
